frozen_requirements.txt
# List project dependencies here. A workflow ports them to `pyproject.toml`. This is needed because Dependabot does not currently support `flit`-flavored `pyproject.toml`.
dvc==2.46.0
loguru==0.6.0
matplotlib==3.7.1
opencv-contrib-python==4.7.0.72
pandas==1.5.3
pandera==0.13.4
pyarrow==11.0.0
pydantic==1.10.6
pyjanitor==0.24.0
pyqtgraph==0.13.2
PySide6==6.4.2
pytz==2022.7.1
PyYAML==6.0
ruamel.yaml==0.17.21
scikit-image==0.20.0
scipy==1.10.1
seaborn==0.12.2
# xarray and its extra dependencies
# See: https://docs.xarray.dev/en/stable/getting-started-guide/installing.html
# Can't specify `accel` extras yet, see: https://github.com/numba/numba/issues/8304
xarray==2023.2.0
Bottleneck==1.3.7
netCDF4==1.6.3
## The following requirements were added by pip freeze:
accessible-pygments==0.0.3
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.13
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.1
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==23.1.0
bleach==6.0.0
blosc2==2.0.0
bokeh==2.4.3
Brotli==1.0.9
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
cftime==1.6.2
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
configobj==5.0.8
contourpy==1.0.7
copier==7.0.1
coverage==7.2.1
cryptography==39.0.2
cycler==0.11.0
Cython==0.29.33
dask==2023.3.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.4.0
distlib==0.3.6
distributed==2023.3.1
distro==1.8.0
docutils==0.19
dpath==2.1.4
dulwich==0.21.3
dunamai==1.16.0
dvc-data==0.42.3
dvc-gs==2.22.0
dvc-http==2.30.2
dvc-objects==0.21.1
dvc-render==0.2.0
dvc-studio-client==0.6.0
dvc-task==0.2.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.1
flatten-dict==0.4.2
flit==3.8.0
flit_core==3.8.0
flufl.lock==7.1.1
fonttools==4.39.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
funcy==1.18
gcsfs==2023.3.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grandalf==0.8
HeapDict==1.0.1
hydra-core==1.3.2
identify==2.5.20
idna==3.4
imageio==2.26.0
imagesize==1.4.1
importlib-metadata==6.0.0
inflate64==0.3.1
iniconfig==2.0.0
ipykernel==6.21.3
ipympl==0.9.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
iteration-utilities==0.11.0
iterative-telemetry==0.0.7
jaraco.classes==3.2.3
jedi==0.18.2
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.13.1
kiwisolver==1.4.4
kombu==5.2.4
lazy_loader==0.1
livereload==2.6.3
locket==1.0.0
lz4==4.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
multipledispatch==0.6.0
multivolumefile==0.2.3
mypy-extensions==1.0.0
myst-parser==1.0.0
nanotime==0.5.2
natsort==8.3.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nbqa==1.6.3
nest-asyncio==1.5.6
networkx==3.0
nodeenv==1.7.0
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.8.7
packaging==23.0
pandas-flavor==0.5.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathspec==0.11.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
plumbum==1.8.1
pre-commit==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.20.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pycine @ git+https://github.com/ottomatic-io/pycine@815cfca06cafc50745a43b2cd0168982225c6dca
pycparser==2.21
pycryptodomex==3.17
pydata-sphinx-theme==0.13.1
pydot==1.4.2
pygit2==1.11.1
Pygments==2.14.0
pygtrie==2.5.0
pyparsing==3.0.9
pyppmd==1.0.0
pyright==1.1.300
pyrsistent==0.19.3
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
pytest==7.2.2
pytest-cov==4.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.10
pyyaml-include==1.3
pyzmq==25.0.1
pyzstd==0.15.4
qtconsole==5.4.1
QtPy==2.3.0
questionary==1.10.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
rsa==4.9
ruff==0.0.254
scmrepo==0.1.15
Send2Trash==1.8.0
shiboken6==6.4.2
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
smmap==5.0.0
snakeviz==2.1.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery-cli==1.0.6
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinx-book-theme==1.0.0
sphinx_design==0.3.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.8.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqltrie==0.3.0
stack-data==0.6.2
tables==3.8.0
tabulate==0.9.0
tblib==1.7.0
terminado==0.17.1
texttable==1.6.7
tifffile==2023.3.15
tinycss2==1.2.1
tokenize-rt==5.0.0
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.11.6
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
voluptuous==0.13.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
win32-setctime==1.1.0
wrapt==1.15.0
yarl==1.8.2
zc.lockfile==3.0.post1
zict==2.2.0
zipp==3.15.0
